+++ CHANGELOG.md
@@ -1,1 +1,5 @@
+# Change Log
 
+## [In Development](https://github.com/ppfeufer/bootstrap-3-shortcodes-for-wordpress)
+### Added
+- Shortcode ```[alert]```